Move load_user_config and add docstring.
