+++ aiaccel/utils/__init__.py
@@ -1,3 +1,3 @@
-from aiaccel.utils.config import pathlib2str_config, print_config
+from aiaccel.utils.config import load_user_config, pathlib2str_config, print_config
 
-__all__ = ["print_config", "pathlib2str_config"]
+__all__ = ["print_config", "pathlib2str_config", "load_user_config"]
