+++ aiaccel/utils/config.py
@@ -33,3 +33,27 @@ def pathlib2str_config(config: ListConfig | DictConfig) -> ListConfig | DictConf
         return config
 
     return _inner_fn(deepcopy(config))
+
+
+def load_user_config(config: Path) -> DictConfig | ListConfig:
+    """Load User Configuration
+
+    When the user specifies _base_, the specified configuration is loaded as the base,
+    and the original configuration is merged with base config.
+    If the configuration specified in _base_ also contains _base_, the process is handled recursively.
+
+    Args:
+        config (Path): Path to the configuration
+
+    Returns:
+        merge_user_config (DictConfig): The merged configuration of the base config and the original config
+        user_config(DictConfig | ListConfig) : The configuration without _base_
+
+    """
+    user_config = oc.load(config)
+    if isinstance(user_config, DictConfig) and "_base_" in user_config:
+        base_config = load_user_config(Path(user_config["_base_"]))
+        merge_user_config = oc.merge(base_config, user_config)
+        return merge_user_config
+    else:
+        return user_config
